backend/requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
python-slugify==8.0.1
email-validator==2.1.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
Flask-Mail==0.9.1
qrcode[pil]==7.4.2
Pillow==10.1.0
